+++ plugins/11_TheWarWithin/localization/esES.lua
@@ -659,6 +659,11 @@ L['george_the_big_pinch_note'] = 'Haz click rápidamente en los 5 {npc:238189} p
 L['options_icons_karesh_lore_hunter'] = '{achievement:60890}'
 L['options_icons_karesh_lore_hunter_desc'] = 'Muestra dónde están los objetos históricos del logro {achievement:60890}.'
 
+L['options_icons_phase_lost_and_found'] = '{achievement:61017}'
+L['options_icons_phase_lost_and_found_desc'] = nil
+L['phaseorb_label'] = nil
+L['phaseorb_note'] = nil
+
 L['options_icons_purrkins'] = '{achievement:42729}'
 L['options_icons_purrkins_desc'] = 'Muestra dónde están los ronroneadores para el logro {achievement:42729}.'
 L['purrkins_note'] = 'Teclea {emote:/acariciar} en el ronroneador.'