+++ plugins/11_TheWarWithin/localization/zhTW.lua
@@ -657,6 +657,11 @@ L['george_the_big_pinch_note'] = nil
 L['options_icons_karesh_lore_hunter'] = '{achievement:60890}'
 L['options_icons_karesh_lore_hunter_desc'] = '顯示成就 {achievement:60890} 所需的知識位置.'
 
+L['options_icons_phase_lost_and_found'] = '{achievement:61017}'
+L['options_icons_phase_lost_and_found_desc'] = nil
+L['phaseorb_label'] = nil
+L['phaseorb_note'] = nil
+
 L['options_icons_purrkins'] = '{achievement:42729}'
 L['options_icons_purrkins_desc'] = nil
 L['purrkins_note'] = nil